Harden Google Sheets integration: remove env token, extract hook, parallelize restore
- Remove VITE_GSHEETS_ACCESS_TOKEN and related env-level secret plumbing
  from vite-env.d.ts, duckdb-context.tsx, and httpfs-extension-controller.ts.
  Per-sheet tokens entered through the wizard are the proper production path.
- Extract business logic from GoogleSheetConfig (373→120 lines) into
  useGSheetConnection hook, following the use-iceberg-connection pattern.
- Cache discoverWorkbook results between Test and Add to avoid redundant
  XLSX fetches when inputs haven't changed.
- Parallelize Google Sheet view restoration in restore.ts using
  Promise.allSettled (secrets still set up sequentially for dedup safety).
- Fix all lint errors: import ordering, max-len, and quote style.
