+++ src/features/duckdb-context/duckdb-context.tsx
@@ -161,7 +161,6 @@ export const DuckDBConnectionPoolProvider = ({
       return '';
     }
   })();
-
   const cleanupWorkerResources = useCallback(() => {
     if (worker.current != null) {
       worker.current.terminate();
