+++ src/store/restore.ts
@@ -70,15 +70,15 @@ import {
   requestFileHandlePersmissions,
 } from '@utils/file-system';
 import { fileSystemService } from '@utils/file-system-adapter';
-import { findUniqueName } from '@utils/helpers';
-import { buildIcebergSecretPayload } from '@utils/iceberg-catalog';
+import { buildGSheetSpreadsheetUrl } from '@utils/gsheet';
 import {
   buildCreateGSheetHttpSecretQuery,
   buildGSheetHttpSecretName,
   resolveGSheetAccessToken,
 } from '@utils/gsheet-auth';
-import { buildGSheetSpreadsheetUrl } from '@utils/gsheet';
 import { shouldResetRestoredScriptQuery } from '@utils/script-query-persistence';
+import { findUniqueName } from '@utils/helpers';
+import { buildIcebergSecretPayload } from '@utils/iceberg-catalog';
 import { getXlsxSheetNames } from '@utils/xlsx';
 import { IDBPDatabase, openDB } from 'idb';
 
@@ -1080,6 +1080,8 @@ export const restoreAppDataFromIDB = async (
   await Promise.all(registerPromises);
 
   // Recreate Google Sheets-backed views that are not associated with local file handles.
+  // Pass 1 (sequential): resolve tokens and create DuckDB HTTP secrets.
+  // Secrets are deduped by secretRef / fileSourceId so each runs at most once.
   const gsheetDataSources = Array.from(dataSources.values()).filter(
     (ds): ds is GSheetSheetView => ds.type === 'gsheet-sheet',
   );
@@ -1088,19 +1090,24 @@ export const restoreAppDataFromIDB = async (
   const warnedMissingGSheetTokens = new Set<string>();
   for (const dataSource of gsheetDataSources) {
     _reservedViews.add(dataSource.viewName);
-    try {
-      if (dataSource.accessMode === 'authorized' && dataSource.secretRef) {
-        const cacheKey = String(dataSource.secretRef);
-        let accessToken = cachedGSheetTokens.get(cacheKey) ?? null;
-        if (!cachedGSheetTokens.has(cacheKey)) {
-          const payload = await getSecret(iDbConn, dataSource.secretRef);
-          accessToken = resolveGSheetAccessToken(payload?.data) ?? null;
-          cachedGSheetTokens.set(cacheKey, accessToken);
-        }
 
-        if (accessToken) {
-          const gsheetHttpSecretName = buildGSheetHttpSecretName(dataSource.fileSourceId);
-          if (!configuredGSheetHttpSecrets.has(gsheetHttpSecretName)) {
+    if (dataSource.accessMode === 'authorized' && dataSource.secretRef) {
+      const cacheKey = String(dataSource.secretRef);
+      if (!cachedGSheetTokens.has(cacheKey)) {
+        try {
+          const payload = await getSecret(iDbConn, dataSource.secretRef);
+          cachedGSheetTokens.set(cacheKey, resolveGSheetAccessToken(payload?.data) ?? null);
+        } catch (error) {
+          cachedGSheetTokens.set(cacheKey, null);
+          console.warn('Failed to read Google Sheet secret:', error);
+        }
+      }
+
+      const accessToken = cachedGSheetTokens.get(cacheKey) ?? null;
+      if (accessToken) {
+        const gsheetHttpSecretName = buildGSheetHttpSecretName(dataSource.fileSourceId);
+        if (!configuredGSheetHttpSecrets.has(gsheetHttpSecretName)) {
+          try {
             await conn.query(
               buildCreateGSheetHttpSecretQuery(
                 gsheetHttpSecretName,
@@ -1109,37 +1116,43 @@ export const restoreAppDataFromIDB = async (
               ),
             );
             configuredGSheetHttpSecrets.add(gsheetHttpSecretName);
-          }
-        } else {
-          if (!warnedMissingGSheetTokens.has(cacheKey)) {
-            warnings.push(
-              `Google Sheet ${dataSource.spreadsheetName} is missing its saved token. Reconnect it in the wizard.`,
-            );
-            warnedMissingGSheetTokens.add(cacheKey);
+          } catch (error) {
+            console.warn('Failed to create Google Sheet HTTP secret:', error);
           }
         }
+      } else if (!warnedMissingGSheetTokens.has(cacheKey)) {
+        warnings.push(
+          `Google Sheet ${dataSource.spreadsheetName} is missing its saved token. Reconnect it in the wizard.`,
+        );
+        warnedMissingGSheetTokens.add(cacheKey);
       }
-
-      const spreadsheetRef =
-        dataSource.spreadsheetUrl || buildGSheetSpreadsheetUrl(dataSource.spreadsheetId);
-      await createGSheetSheetView(
-        conn,
-        spreadsheetRef,
-        dataSource.sheetName,
-        dataSource.viewName,
-        dataSource.accessMode,
-      );
-      validDataSources.add(dataSource.id);
-    } catch (error) {
-      // Keep persisted Google Sheets data sources on transient restore failures
-      // (offline/expired token/API hiccups) so users can reconnect later.
-      validDataSources.add(dataSource.id);
-      warnings.push(
-        `Google Sheet ${dataSource.spreadsheetName}::${dataSource.sheetName} could not be restored. It will remain saved and can be retried later.`,
-      );
-      console.warn('Failed to restore Google Sheet view:', error);
     }
   }
+
+  // Pass 2 (parallel): create views for each sheet concurrently.
+  await Promise.allSettled(
+    gsheetDataSources.map(async (dataSource) => {
+      try {
+        const spreadsheetRef =
+          dataSource.spreadsheetUrl || buildGSheetSpreadsheetUrl(dataSource.spreadsheetId);
+        await createGSheetSheetView(
+          conn,
+          spreadsheetRef,
+          dataSource.sheetName,
+          dataSource.viewName,
+          dataSource.accessMode,
+        );
+      } catch (error) {
+        // Keep persisted Google Sheets data sources on transient restore failures
+        // (offline/expired token/API hiccups) so users can reconnect later.
+        warnings.push(
+          `Google Sheet ${dataSource.spreadsheetName}::${dataSource.sheetName} could not be restored. It will remain saved and can be retried later.`,
+        );
+        console.warn('Failed to restore Google Sheet view:', error);
+      }
+      validDataSources.add(dataSource.id);
+    }),
+  );
 
   // Handle remote databases and iceberg catalogs - they need to be re-attached.
   // We don't re-attach them here because:
